+++ src/GafferScene/CollectScenes.cpp
@@ -41,11 +41,135 @@
 #include "Gaffer/Context.h"
 #include "Gaffer/StringPlug.h"
 
+#include "IECore/NullObject.h"
+
+#include "boost/container/flat_map.hpp"
+
 using namespace std;
 using namespace Imath;
 using namespace IECore;
 using namespace Gaffer;
 using namespace GafferScene;
+
+//////////////////////////////////////////////////////////////////////////
+// RootTree
+//////////////////////////////////////////////////////////////////////////
+
+namespace
+{
+
+// Used to represent the tree of locations specified by `rootNamesPlug()`.
+// This allows us to quickly find the matching root for any output location
+// via `locationOrAncestor()`.
+class RootTree : public IECore::Data
+{
+
+	public :
+
+		struct Location;
+		using LocationPtr = std::unique_ptr<Location>;
+		using ChildMap = boost::container::flat_map<IECore::InternedString, LocationPtr>;
+
+		struct Location
+		{
+
+			Location( size_t depth )
+				:	depth( depth ), childNames( new InternedStringVectorData() )
+			{
+			}
+
+			bool isRoot() const
+			{
+				return !rootVariableValue.empty();
+			}
+
+			// The path to this location, but exactly as the user spelled
+			// it in `rootNamesPlug()` (may or may not have a leading or
+			// trailing '/'). Empty if not a collection root.
+			string rootVariableValue;
+			size_t depth;
+			ChildMap children;
+			InternedStringVectorDataPtr childNames;
+
+		};
+
+		RootTree( const IECore::StringVectorData *roots )
+			:	m_treeRoot( new Location( 0 ) )
+		{
+			ScenePlug::ScenePath path;
+			for( const auto &root : roots->readable() )
+			{
+				ScenePlug::stringToPath( root, path );
+				if( path.empty() )
+				{
+					continue;
+				}
+
+				Location *location = m_treeRoot.get();
+				for( const auto &name : path )
+				{
+					const auto inserted = location->children.insert( ChildMap::value_type( name, LocationPtr() ) );
+					if( inserted.second )
+					{
+						if( location->isRoot() )
+						{
+							throw IECore::Exception( boost::str( boost::format( "\"%1%\" contains nested roots" ) % location->rootVariableValue ) );
+						}
+						inserted.first->second.reset( new Location( location->depth + 1 ) );
+						location->childNames->writable().push_back( name );
+					}
+					location = inserted.first->second.get();
+				}
+
+				if( location->isRoot() )
+				{
+					// Duplicate found - skip.
+					continue;
+				}
+
+				if( !location->children.empty() )
+				{
+					throw IECore::Exception( boost::str( boost::format( "\"%1%\" contains nested roots" ) % root ) );
+				}
+
+				location->rootVariableValue = root;
+				m_roots.push_back( root );
+			}
+		}
+
+		const Location *locationOrAncestor( const ScenePlug::ScenePath &path ) const
+		{
+			const Location *result = m_treeRoot.get();
+			for( const auto &name : path )
+			{
+				const auto it = result->children.find( name );
+				if( it != result->children.end() )
+				{
+					result = it->second.get();
+				}
+				else
+				{
+					break;
+				}
+			}
+			return result;
+		}
+
+		const vector<string> &roots() const
+		{
+			return m_roots;
+		}
+
+	private :
+
+		LocationPtr m_treeRoot;
+		vector<string> m_roots;
+
+};
+
+IE_CORE_DECLAREPTR( RootTree )
+
+} // namespace
 
 //////////////////////////////////////////////////////////////////////////
 // SourceScope and SourcePathScope
@@ -56,19 +180,19 @@ class CollectScenes::SourceScope : public Context::EditableScope
 
 	public :
 
-		SourceScope( const Context *context, const InternedString &rootNameVariable )
-			:	EditableScope( context ), m_rootNameVariable( rootNameVariable )
+		SourceScope( const Context *context, const InternedString &rootVariable )
+			:	EditableScope( context ), m_rootVariable( rootVariable )
 		{
 		}
 
-		void setRootName( const InternedString &name )
+		void setRoot( const std::string &root )
 		{
-			set( m_rootNameVariable, name.string() );
+			set( m_rootVariable, root );
 		}
 
 	private :
 
-		InternedString m_rootNameVariable;
+		InternedString m_rootVariable;
 
 };
 
@@ -80,23 +204,45 @@ class CollectScenes::SourcePathScope : public SourceScope
 		SourcePathScope( const Context *context, const CollectScenes *collectScenes, const ScenePlug::ScenePath &downstreamPath )
 			:	SourceScope( context, collectScenes->rootNameVariablePlug()->getValue() )
 		{
-			setRootName( downstreamPath[0] );
-			ScenePlug::ScenePath upstreamPath;
-			// We evaluate the sourceRootPlug _after_ setting the root name,
-			// so that users can use the root name in expressions and
-			// substitutions.
-			ScenePlug::stringToPath( collectScenes->sourceRootPlug()->getValue(), upstreamPath );
-			upstreamPath.insert( upstreamPath.end(), downstreamPath.begin() + 1, downstreamPath.end() );
-			set( ScenePlug::scenePathContextName, upstreamPath );
+			// Evaluate RootTree in global scope.
+			remove( ScenePlug::scenePathContextName );
+			m_rootTree = boost::static_pointer_cast<const RootTree>( collectScenes->rootTreePlug()->getValue() );
+			m_rootTreeLocation = m_rootTree->locationOrAncestor( downstreamPath );
+			if( m_rootTreeLocation->isRoot() )
+			{
+				setRoot( m_rootTreeLocation->rootVariableValue );
+				ScenePlug::ScenePath upstreamPath;
+				// We evaluate the sourceRootPlug _after_ setting the root name,
+				// so that users can use the root name in expressions and
+				// substitutions.
+				ScenePlug::stringToPath( collectScenes->sourceRootPlug()->getValue(), upstreamPath );
+				upstreamPath.insert( upstreamPath.end(), downstreamPath.begin() + m_rootTreeLocation->depth, downstreamPath.end() );
+				set( ScenePlug::scenePathContextName, upstreamPath );
+			}
+			else
+			{
+				set( ScenePlug::scenePathContextName, downstreamPath );
+			}
+		}
+
+		const RootTree::Location *rootTreeLocation() const
+		{
+			return m_rootTreeLocation;
 		}
 
 		static bool affectedBy( const CollectScenes *collectScenes, const Plug *input )
 		{
 			return
 				input == collectScenes->rootNameVariablePlug() ||
+				input == collectScenes->rootTreePlug() ||
 				input == collectScenes->sourceRootPlug()
 			;
 		}
+
+	private :
+
+		ConstRootTreePtr m_rootTree;
+		const RootTree::Location *m_rootTreeLocation;
 
 };
 
@@ -116,6 +262,7 @@ CollectScenes::CollectScenes( const std::string &name )
 	addChild( new StringVectorDataPlug( "rootNames", Plug::In, new StringVectorData() ) );
 	addChild( new StringPlug( "rootNameVariable", Plug::In, "collect:rootName" ) );
 	addChild( new StringPlug( "sourceRoot", Plug::In, "/" ) );
+	addChild( new ObjectPlug( "__rootTree", Plug::Out, IECore::NullObject::defaultNullObject() ) );
 
 	outPlug()->childBoundsPlug()->setFlags( Plug::AcceptsDependencyCycles, true );
 }
@@ -154,9 +301,24 @@ const Gaffer::StringPlug *CollectScenes::sourceRootPlug() const
 	return getChild<StringPlug>( g_firstPlugIndex + 2 );
 }
 
+Gaffer::ObjectPlug *CollectScenes::rootTreePlug()
+{
+	return getChild<ObjectPlug>( g_firstPlugIndex + 3 );
+}
+
+const Gaffer::ObjectPlug *CollectScenes::rootTreePlug() const
+{
+	return getChild<ObjectPlug>( g_firstPlugIndex + 3 );
+}
+
 void CollectScenes::affects( const Gaffer::Plug *input, AffectedPlugsContainer &outputs ) const
 {
 	SceneProcessor::affects( input, outputs );
+
+	if( input == rootNamesPlug() )
+	{
+		outputs.push_back( rootTreePlug() );
+	}
 
 	if(
 		input == outPlug()->childBoundsPlug() ||
@@ -183,7 +345,6 @@ void CollectScenes::affects( const Gaffer::Plug *input, AffectedPlugsContainer &
 	}
 
 	if(
-		input == rootNamesPlug() ||
 		SourcePathScope::affectedBy( this, input ) ||
 		input == inPlug()->existsPlug() ||
 		input == inPlug()->childNamesPlug()
@@ -193,7 +354,7 @@ void CollectScenes::affects( const Gaffer::Plug *input, AffectedPlugsContainer &
 	}
 
 	if(
-		input == outPlug()->childNamesPlug() ||
+		input == rootTreePlug() ||
 		input == rootNameVariablePlug() ||
 		input == inPlug()->globalsPlug()
 	)
@@ -202,7 +363,7 @@ void CollectScenes::affects( const Gaffer::Plug *input, AffectedPlugsContainer &
 	}
 
 	if(
-		input == outPlug()->childNamesPlug() ||
+		input == rootTreePlug() ||
 		input == inPlug()->setNamesPlug() ||
 		input == rootNameVariablePlug()
 	)
@@ -211,7 +372,7 @@ void CollectScenes::affects( const Gaffer::Plug *input, AffectedPlugsContainer &
 	}
 
 	if(
-		input == outPlug()->childNamesPlug() ||
+		input == rootTreePlug() ||
 		input == inPlug()->setPlug() ||
 		input == sourceRootPlug() ||
 		input == rootNameVariablePlug()
@@ -221,124 +382,146 @@ void CollectScenes::affects( const Gaffer::Plug *input, AffectedPlugsContainer &
 	}
 }
 
+void CollectScenes::hash( const Gaffer::ValuePlug *output, const Gaffer::Context *context, IECore::MurmurHash &h ) const
+{
+	SceneProcessor::hash( output, context, h );
+
+	if( output == rootTreePlug() )
+	{
+		rootNamesPlug()->hash( h );
+	}
+}
+
+void CollectScenes::compute( Gaffer::ValuePlug *output, const Gaffer::Context *context ) const
+{
+	if( output == rootTreePlug() )
+	{
+		ConstStringVectorDataPtr roots = rootNamesPlug()->getValue();
+		static_cast<ObjectPlug *>( output )->setValue(
+			new RootTree( roots.get() )
+		);
+		return;
+	}
+
+	SceneProcessor::compute( output, context );
+}
+
 void CollectScenes::hashBound( const ScenePath &path, const Gaffer::Context *context, const ScenePlug *parent, IECore::MurmurHash &h ) const
 {
-	if( path.size() == 0 )
+	SourcePathScope sourcePathScope( context, this, path );
+	if( !sourcePathScope.rootTreeLocation()->isRoot() )
 	{
 		h = outPlug()->childBoundsPlug()->hash();
 	}
 	else
 	{
-		SourcePathScope sourcePathScope( context, this, path );
 		h = inPlug()->boundPlug()->hash();
 	}
 }
 
 Imath::Box3f CollectScenes::computeBound( const ScenePath &path, const Gaffer::Context *context, const ScenePlug *parent ) const
 {
-	if( path.size() == 0 )
+	SourcePathScope sourcePathScope( context, this, path );
+	if( !sourcePathScope.rootTreeLocation()->isRoot() )
 	{
 		return outPlug()->childBoundsPlug()->getValue();
 	}
 	else
 	{
-		SourcePathScope sourcePathScope( context, this, path );
 		return inPlug()->boundPlug()->getValue();
 	}
 }
 
 void CollectScenes::hashTransform( const ScenePath &path, const Gaffer::Context *context, const ScenePlug *parent, IECore::MurmurHash &h ) const
 {
-	if( path.size() == 0 )
+	SourcePathScope sourcePathScope( context, this, path );
+	if( !sourcePathScope.rootTreeLocation()->isRoot() )
 	{
 		SceneProcessor::hashTransform( path, context, parent, h );
 	}
 	else
 	{
-		SourcePathScope sourcePathScope( context, this, path );
 		h = inPlug()->transformPlug()->hash();
 	}
 }
 
 Imath::M44f CollectScenes::computeTransform( const ScenePath &path, const Gaffer::Context *context, const ScenePlug *parent ) const
 {
-	if( path.size() == 0 )
+	SourcePathScope sourcePathScope( context, this, path );
+	if( !sourcePathScope.rootTreeLocation()->isRoot() )
 	{
 		return M44f();
 	}
 	else
 	{
-		SourcePathScope sourcePathScope( context, this, path );
 		return inPlug()->transformPlug()->getValue();
 	}
 }
 
 void CollectScenes::hashAttributes( const ScenePath &path, const Gaffer::Context *context, const ScenePlug *parent, IECore::MurmurHash &h ) const
 {
-	if( path.size() == 0 )
+	SourcePathScope sourcePathScope( context, this, path );
+	if( !sourcePathScope.rootTreeLocation()->isRoot() )
 	{
 		SceneProcessor::hashAttributes( path, context, parent, h );
 	}
 	else
 	{
-		SourcePathScope sourcePathScope( context, this, path );
 		h = inPlug()->attributesPlug()->hash();
 	}
 }
 
 IECore::ConstCompoundObjectPtr CollectScenes::computeAttributes( const ScenePath &path, const Gaffer::Context *context, const ScenePlug *parent ) const
 {
-	if( path.size() == 0 )
+	SourcePathScope sourcePathScope( context, this, path );
+	if( !sourcePathScope.rootTreeLocation()->isRoot() )
 	{
 		return outPlug()->attributesPlug()->defaultValue();
 	}
 	else
 	{
-		SourcePathScope sourcePathScope( context, this, path );
 		return inPlug()->attributesPlug()->getValue();
 	}
 }
 
 void CollectScenes::hashObject( const ScenePath &path, const Gaffer::Context *context, const ScenePlug *parent, IECore::MurmurHash &h ) const
 {
-	if( path.size() == 0 )
+	SourcePathScope sourcePathScope( context, this, path );
+	if( !sourcePathScope.rootTreeLocation()->isRoot() )
 	{
 		SceneProcessor::hashObject( path, context, parent, h );
 	}
 	else
 	{
-		SourcePathScope sourcePathScope( context, this, path );
 		h = inPlug()->objectPlug()->hash();
 	}
 }
 
 IECore::ConstObjectPtr CollectScenes::computeObject( const ScenePath &path, const Gaffer::Context *context, const ScenePlug *parent ) const
 {
-	if( path.size() == 0 )
+	SourcePathScope sourcePathScope( context, this, path );
+	if( !sourcePathScope.rootTreeLocation()->isRoot() )
 	{
 		return outPlug()->objectPlug()->defaultValue();
 	}
 	else
 	{
-		SourcePathScope sourcePathScope( context, this, path );
 		return inPlug()->objectPlug()->getValue();
 	}
 }
 
 void CollectScenes::hashChildNames( const ScenePath &path, const Gaffer::Context *context, const ScenePlug *parent, IECore::MurmurHash &h ) const
 {
-	if( path.size() == 0 )
+	SourcePathScope sourcePathScope( context, this, path );
+	if( !sourcePathScope.rootTreeLocation()->isRoot() )
 	{
-		SceneProcessor::hashChildNames( path, context, parent, h );
-		rootNamesPlug()->hash( h );
+		h = sourcePathScope.rootTreeLocation()->childNames->Object::hash();
 	}
 	else
 	{
-		SourcePathScope sourcePathScope( context, this, path );
-		if( path.size() == 1 )
+		if( path.size() == sourcePathScope.rootTreeLocation()->depth )
 		{
-			const auto &upstreamPath = Context::current()->get<ScenePlug::ScenePath>( ScenePlug::scenePathContextName );
-			if( !inPlug()->exists( upstreamPath ) )
+			if( !inPlug()->existsPlug()->getValue() )
 			{
 				h = inPlug()->childNamesPlug()->defaultValue()->Object::hash();
 				return;
@@ -350,36 +533,16 @@ void CollectScenes::hashChildNames( const ScenePath &path, const Gaffer::Context
 
 IECore::ConstInternedStringVectorDataPtr CollectScenes::computeChildNames( const ScenePath &path, const Gaffer::Context *context, const ScenePlug *parent ) const
 {
-	if( path.size() == 0 )
+	SourcePathScope sourcePathScope( context, this, path );
+	if( !sourcePathScope.rootTreeLocation()->isRoot() )
 	{
-		ConstStringVectorDataPtr rootNamesData = rootNamesPlug()->getValue();
-		const vector<string> &rootNames = rootNamesData->readable();
-
-		InternedStringVectorDataPtr childNamesData = new InternedStringVectorData;
-		vector<InternedString> &childNames = childNamesData->writable();
-
-		for( vector<string>::const_iterator it = rootNames.begin(), eIt = rootNames.end(); it != eIt; ++it )
-		{
-			if( it->empty() )
-			{
-				continue;
-			}
-			InternedString childName( *it );
-			if( find( childNames.begin(), childNames.end(), childName ) == childNames.end() )
-			{
-				childNames.push_back( childName );
-			}
-		}
-
-		return childNamesData;
+		return sourcePathScope.rootTreeLocation()->childNames;
 	}
 	else
 	{
-		SourcePathScope sourcePathScope( context, this, path );
-		if( path.size() == 1 )
+		if( path.size() == sourcePathScope.rootTreeLocation()->depth )
 		{
-			const auto &upstreamPath = Context::current()->get<ScenePlug::ScenePath>( ScenePlug::scenePathContextName );
-			if( !inPlug()->exists( upstreamPath ) )
+			if( !inPlug()->existsPlug()->getValue() )
 			{
 				return inPlug()->childNamesPlug()->defaultValue();
 			}
@@ -390,12 +553,11 @@ IECore::ConstInternedStringVectorDataPtr CollectScenes::computeChildNames( const
 
 void CollectScenes::hashGlobals( const Gaffer::Context *context, const ScenePlug *parent, IECore::MurmurHash &h ) const
 {
-	ConstInternedStringVectorDataPtr rootNamesData = outPlug()->childNames( ScenePath() );
-	const vector<InternedString> &rootNames = rootNamesData->readable();
-	if( rootNames.size() )
+	ConstRootTreePtr rootTree = boost::static_pointer_cast<const RootTree>( rootTreePlug()->getValue() );
+	if( rootTree->roots().size() )
 	{
 		SourceScope sourceScope( context, rootNameVariablePlug()->getValue() );
-		sourceScope.setRootName( rootNames[0] );
+		sourceScope.setRoot( rootTree->roots()[0] );
 		h = inPlug()->globalsPlug()->hash();
 	}
 	else
@@ -406,12 +568,11 @@ void CollectScenes::hashGlobals( const Gaffer::Context *context, const ScenePlug
 
 IECore::ConstCompoundObjectPtr CollectScenes::computeGlobals( const Gaffer::Context *context, const ScenePlug *parent ) const
 {
-	ConstInternedStringVectorDataPtr rootNamesData = outPlug()->childNames( ScenePath() );
-	const vector<InternedString> &rootNames = rootNamesData->readable();
-	if( rootNames.size() )
+	ConstRootTreePtr rootTree = boost::static_pointer_cast<const RootTree>( rootTreePlug()->getValue() );
+	if( rootTree->roots().size() )
 	{
 		SourceScope sourceScope( context, rootNameVariablePlug()->getValue() );
-		sourceScope.setRootName( rootNames[0] );
+		sourceScope.setRoot( rootTree->roots()[0] );
 		return inPlug()->globalsPlug()->getValue();
 	}
 	else
@@ -424,23 +585,20 @@ void CollectScenes::hashSetNames( const Gaffer::Context *context, const ScenePlu
 {
 	SceneProcessor::hashSetNames( context, parent, h );
 
-	ConstInternedStringVectorDataPtr rootNamesData = outPlug()->childNames( ScenePath() );
-	const vector<InternedString> &rootNames = rootNamesData->readable();
-
+	ConstRootTreePtr rootTree = boost::static_pointer_cast<const RootTree>( rootTreePlug()->getValue() );
 	const ValuePlug *inSetNamesPlug = inPlug()->setNamesPlug();
 
 	SourceScope sourceScope( context, rootNameVariablePlug()->getValue() );
-	for( vector<InternedString>::const_iterator it = rootNames.begin(), eIt = rootNames.end(); it != eIt; ++it )
+	for( const auto &root : rootTree->roots() )
 	{
-		sourceScope.setRootName( *it );
+		sourceScope.setRoot( root );
 		inSetNamesPlug->hash( h );
 	}
 }
 
 IECore::ConstInternedStringVectorDataPtr CollectScenes::computeSetNames( const Gaffer::Context *context, const ScenePlug *parent ) const
 {
-	ConstInternedStringVectorDataPtr rootNamesData = outPlug()->childNames( ScenePath() );
-	const vector<InternedString> &rootNames = rootNamesData->readable();
+	ConstRootTreePtr rootTree = boost::static_pointer_cast<const RootTree>( rootTreePlug()->getValue() );
 
 	InternedStringVectorDataPtr setNamesData = new InternedStringVectorData;
 	vector<InternedString> &setNames = setNamesData->writable();
@@ -448,16 +606,15 @@ IECore::ConstInternedStringVectorDataPtr CollectScenes::computeSetNames( const G
 	const InternedStringVectorDataPlug *inSetNamesPlug = inPlug()->setNamesPlug();
 
 	SourceScope sourceScope( context, rootNameVariablePlug()->getValue() );
-	for( vector<InternedString>::const_iterator it = rootNames.begin(), eIt = rootNames.end(); it != eIt; ++it )
+	for( const auto &root : rootTree->roots() )
 	{
-		sourceScope.setRootName( *it );
+		sourceScope.setRoot( root );
 		ConstInternedStringVectorDataPtr inSetNamesData = inSetNamesPlug->getValue();
-		const vector<InternedString> &inSetNames = inSetNamesData->readable();
-		for( vector<InternedString>::const_iterator sIt = inSetNames.begin(), sEIt = inSetNames.end(); sIt != sEIt; ++sIt )
+		for( const auto &setName : inSetNamesData->readable() )
 		{
-			if( find( setNames.begin(), setNames.end(), *sIt ) == setNames.end() )
+			if( find( setNames.begin(), setNames.end(), setName ) == setNames.end() )
 			{
-				setNames.push_back( *sIt );
+				setNames.push_back( setName );
 			}
 		}
 	}
@@ -469,43 +626,49 @@ void CollectScenes::hashSet( const IECore::InternedString &setName, const Gaffer
 {
 	SceneProcessor::hashSet( setName, context, parent, h );
 
-	ConstInternedStringVectorDataPtr rootNamesData = outPlug()->childNames( ScenePath() );
-	const vector<InternedString> &rootNames = rootNamesData->readable();
+	ConstRootTreePtr rootTree;
+	{
+		ScenePlug::GlobalScope globalScope( context );
+		rootTree = boost::static_pointer_cast<const RootTree>( rootTreePlug()->getValue() );
+	}
 
 	const PathMatcherDataPlug *inSetPlug = inPlug()->setPlug();
 	const StringPlug *sourceRootPlug = this->sourceRootPlug();
 
 	SourceScope sourceScope( context, rootNameVariablePlug()->getValue() );
-	for( vector<InternedString>::const_iterator it = rootNames.begin(), eIt = rootNames.end(); it != eIt; ++it )
+	for( const auto &root : rootTree->roots() )
 	{
-		sourceScope.setRootName( *it );
+		sourceScope.setRoot( root );
 		inSetPlug->hash( h );
 		sourceRootPlug->hash( h );
-		h.append( *it );
+		h.append( root );
 	}
 }
 
 IECore::ConstPathMatcherDataPtr CollectScenes::computeSet( const IECore::InternedString &setName, const Gaffer::Context *context, const ScenePlug *parent ) const
 {
+	ConstRootTreePtr rootTree;
+	{
+		ScenePlug::GlobalScope globalScope( context );
+		rootTree = boost::static_pointer_cast<const RootTree>( rootTreePlug()->getValue() );
+	}
+
 	PathMatcherDataPtr setData = new PathMatcherData;
 	PathMatcher &set = setData->writable();
-
-	ConstInternedStringVectorDataPtr rootNamesData = outPlug()->childNames( ScenePath() );
-	const vector<InternedString> &rootNames = rootNamesData->readable();
 
 	const PathMatcherDataPlug *inSetPlug = inPlug()->setPlug();
 	const StringPlug *sourceRootPlug = this->sourceRootPlug();
 
 	SourceScope sourceScope( context, rootNameVariablePlug()->getValue() );
-	vector<InternedString> prefix( 1 );
-	for( vector<InternedString>::const_iterator it = rootNames.begin(), eIt = rootNames.end(); it != eIt; ++it )
+	ScenePlug::ScenePath prefix;
+	for( const auto &root : rootTree->roots() )
 	{
-		sourceScope.setRootName( *it );
+		sourceScope.setRoot( root );
 		ConstPathMatcherDataPtr inSetData = inSetPlug->getValue();
 		const PathMatcher &inSet = inSetData->readable();
 		if( !inSet.isEmpty() )
 		{
-			prefix[0] = *it;
+			ScenePlug::stringToPath( root, prefix );
 			const string root = sourceRootPlug->getValue();
 			if( !root.empty() )
 			{