+++ Changes.md
@@ -4,6 +4,7 @@
 Improvements
 ------------
 
+- CollectScenes : Added support for collecting at roots of arbitrary depth.
 - AttributeVisualiser : Added support for V2f, V3f, V2i, V3i, V2d and V3d data types.
 
 Fixes